좋아요컴포턴트분리
