+++ src/components/Lweet.js
@@ -1,18 +1,15 @@
 import React from 'react';
 import { db,storage } from "fbase.js";
-import { doc, collection,addDoc,deleteDoc,  updateDoc, onSnapshot,getDocs } from "firebase/firestore";
+import { doc,deleteDoc,  updateDoc } from "firebase/firestore";
 import { deleteObject, ref } from "firebase/storage"
-import { useState, useEffect } from "react";
-//import Like from "components/Like.js";
+import { useState } from "react";
+
 
 const Lweet = ({ lweetObj, isOwner, userObj }) => {
    
     const [editing, setEditing] = useState(false);
     const [newLweet, setNewLweet] = useState(lweetObj.text);
-    //const [ like, setLike]= useState([]);
-    const [ likes, setLikes]= useState([]);
-    //const [ liketotal, setliketotal] =  useState(lweetObj.total);
-    const [ likemg, setLikemg]= useState(0);
+  
 
     
 
@@ -50,25 +47,7 @@ const Lweet = ({ lweetObj, isOwner, userObj }) => {
         //await db.doc(`lweets/${lweetObj.id}`).update({text: newLweet});
     };
 
-    const onLike = async(event) => {
-        event.preventDefault();
-        const ok = window.confirm("좋아요를 누르시겠습니까?");
-        console.log('%suserObj',userObj);
-        if (ok){
-            const post = doc(db , `lweets/${lweetObj.id}`);
-            console.log('%post',post.id);
-            await addDoc(collection(db, "likes"), {
-            post:post.id,
-            uid:userObj.uid,
-            createdAt: Date.now(),
-        })
-
-        // if(userObj.uid === doc(db , `likes/${lweetObj.id}`);) 
-       
-    };
-    //updateDoc(doc(db, `lweets/${lweetObj.id}`), {total: liketotal});
-    setLikemg(1);
-}
+  
 
     // const Likes = async() => {
     //     const dbLweets = await getDocs(collection(db, "likes"));
@@ -90,22 +69,7 @@ const Lweet = ({ lweetObj, isOwner, userObj }) => {
       
     // }, []);
 
-        useEffect(() => {
-        onSnapshot(collection(db,"likes"), (snapshot) =>
-        {
-            const newLike = snapshot.docs.map((document) => ({
-                id: document.id,
-                uid: userObj.uid,
-               ...document.data(),
-              
-                
-            }));
-            console.log('%s 라이크' ,newLike.uid);
-            setLikes(newLike);
-        });
       
-       
-    }, []);
     
   
 
@@ -116,19 +80,7 @@ const Lweet = ({ lweetObj, isOwner, userObj }) => {
 //     </li>
 //   );
 
-    const togglelike = async() => {
-        const ok = window.confirm("좋아요 취소?");
-      
-        if (ok) {
-            console.log('%s 아이디2',likes.map((like) => like.id));
-            const LikeData = likes.map((like)  => <li likeonj = {like}  key ={ like.id} /> )
-            console.log('%s LikeData',LikeData.likeonj);
-           const data = await deleteDoc(doc(db , `likes/${LikeData.id}`));
-           //const test = await doc(db,`likes/${likeID}`);
-            console.log('%s data',data);
-        };
-        setLikemg((prev) => !prev);
-    }
+   
 
     return(
         <div className='body'>
@@ -156,14 +108,7 @@ const Lweet = ({ lweetObj, isOwner, userObj }) => {
             <button onClick = {toggleEditing}>Update Lweet </button>
             </>
             )
-        }
-                { likemg ? ( 
-              <button onClick = {togglelike}> Lweet Cancel </button>
-                ):( 
-                    <button onClick = {onLike}> Lweet like </button>
-               
-                )
-                }
+        }
