+++ src/routes/Home.js
@@ -5,7 +5,7 @@ import { collection, onSnapshot, getDocs } from "firebase/firestore";
 import Lweet from "components/Lweet.js";
 import LweetFactory from 'components/LweetFactory.js';
 // import AppRouter from 'components/Router.js';
-// import Like from "components/Like.js";
+import Like from "components/Like.js";
 
 
 
@@ -13,24 +13,7 @@ const Home = ({userObj}) => {
     
    
     const [ lweets, setLweets ] = useState([]);
-    //const [ likes, setLikes ] = useState([]);
-   
-
-
-    // const Likes = async() => {
-    //     const dbLweets = await getDocs(collection(db, "likes"));
-    //     dbLweets.forEach((document) => {
-    //         console.log('%s 이거는',document.data)
-    //         //const lweetObject = { ...document.data(), id: document.id };
-    //       //  setLikes((prev) => [lweetObject, ...prev])
-    //        // console.log('%s like',lweetObject);
-    // });
-    // useEffect(() => {
-    //     Likes();
-    // }, []);
-     
-    // }
-    // console.log(likes);
+    const [ likes, setLikes]= useState([]);
 
 
 
@@ -47,41 +30,30 @@ const Home = ({userObj}) => {
       
      
     }, []);
-   
-    // useEffect(() => {
-    //     onSnapshot(collection(db,"likes"), (snapshot) =>
-    //     {
-    //         const newLike = snapshot.docs.map((document) => ({
-    //             id: document.id,
-    //             ...document.data(),
-    //         }));
-    //         console.log('%s 라이크' ,newLike.id);
-    //         setLikes(newLike);
-    //     });
-     
-    // }, []);
-
- 
+    console.log("%s오류", lweets);
     return (
         <>
-       
+        
     <LweetFactory userObj={userObj} />
-    {/* <div> {likes.map((like)=> (
-        <Lweet 
-        key = {like.id}
-        likeobj = {like}/>
-    ))}</div> */}
-
              <div>
-                {lweets.map((lweet) => (
-
+                {lweets.map((lweet, id) => (
+                    <>
                     <Lweet 
-                    key = { lweet.id} 
+                    //key = {id} 
                     lweetObj={lweet}
-                    userObj={userObj}
                     isOwner={lweet.creatorID === userObj.uid} 
                     />   
+                    
+                    
+                    <Like 
+                    
+                    lweetObj={lweet}
+                    userObj={userObj}
+                    /> 
+                
+                    </>
                 ))}
+                
              </div>
 
              </>
